refactor: merge the qmp client into a single QemuMonitorClient
Fold the QMP wire protocol into the type that uses it. The monitor is
now one type with four public methods where it was two types exposing
nine. Sending, receiving and command dispatch become private.

Drop the verbosity field. Every construction site passed Normal, so
the two trace prints it guarded could never run.
