+++ src/qemu.rs
@@ -1,18 +1,16 @@
-mod monitor;
 mod qemu_firmware;
 mod qemu_firmware_descriptor;
 mod qemu_img;
+mod qemu_monitor_client;
 mod qemu_path_builder;
 mod qemu_system;
-mod qmp;
 mod qmp_message;
 mod tls_client;
 
-pub use monitor::*;
 pub use qemu_firmware::{QemuFirmware, QemuInstall};
 pub use qemu_img::*;
+pub use qemu_monitor_client::*;
 pub use qemu_path_builder::QemuPathBuilder;
 pub use qemu_system::*;
-pub use qmp::*;
 pub use qmp_message::*;
 pub use tls_client::*;
